Change index2.js to index.js
Deleted original index.js
Moved auth to short polling. Makes it easier on the client side when handling promises. Also makes it more annoying when you aren't logged in because of the 1 second delay.
