+++ index.html
@@ -244,6 +244,6 @@
     </div>
     <!-- If code loading times matter, comment out the next line, and change the one after that to 'index2.js' -->
     <!-- <script src="https://code.jquery.com/jquery-3.7.0.min.js" integrity="sha256-2Pmvv0kuTBOenSvLm6bvfBSSHrUJ+3A7x6P5Ebd07/g=" crossorigin="anonymous"></script> -->
-    <script src="js/index2.js"></script>
+    <script src="js/index.js"></script>
   </body>
 </html>
